refactoring : intellij refactoring format, contentToTask abstract class에서 제거
