+++ app/src/main/java/com/codesoom/assignment/method/DoTask.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.codesoom.assignment.models.Task;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sun.net.httpserver.HttpExchange;
 
@@ -23,11 +22,12 @@ public abstract class DoTask {
 
 	public abstract void handleItem() throws IOException;
 
-	protected abstract Task contentToTask(String content) throws JsonProcessingException;
-
 	protected String taskToJSON(Long id) throws IOException {
 		OutputStream outputStream = new ByteArrayOutputStream();
-		Task task = tasks.stream().filter(t -> t.getId().equals(id)).findAny().orElseThrow(IllegalArgumentException::new);
+		Task task = tasks.stream()
+			.filter(t -> t.getId().equals(id))
+			.findAny()
+			.orElseThrow(IllegalArgumentException::new);
 		objectMapper.writeValue(outputStream, task);
 		return outputStream.toString();
 	}
@@ -40,6 +40,15 @@ public abstract class DoTask {
 
 	protected Long getIdFromPath(HttpExchange exchange) {
 		String idString = exchange.getRequestURI().getPath().substring("/tasks/".length());
+		canConvertId(idString);
 		return Long.valueOf(idString);
+	}
+
+	private void canConvertId(String idString) {
+		try {
+			Long.valueOf(idString);
+		} catch (Exception e) {
+			throw new IllegalArgumentException();
+		}
 	}
 }
