+++ app/src/main/java/com/codesoom/assignment/method/GetAllTask.java
@@ -2,9 +2,7 @@ package com.codesoom.assignment.method;
 
 import java.io.IOException;
 
-import com.codesoom.assignment.models.Task;
 import com.codesoom.assignment.response.ResponseOK;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.sun.net.httpserver.HttpExchange;
 
 public class GetAllTask extends DoTask {
@@ -14,12 +12,6 @@ public class GetAllTask extends DoTask {
 
 	@Override
 	public void handleItem() throws IOException {
-		String content = tasksToJSON();
-		new ResponseOK(exchange).send(content);
-	}
-
-	@Override
-	protected Task contentToTask(String content) throws JsonProcessingException {
-		return null;
+		new ResponseOK(exchange).send(tasksToJSON());
 	}
 }
