+++ app/src/main/java/com/codesoom/assignment/response/Response.java
@@ -2,15 +2,12 @@ package com.codesoom.assignment.response;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.ArrayList;
-import java.util.List;
 
-import com.codesoom.assignment.models.Task;
 import com.sun.net.httpserver.HttpExchange;
 
 public abstract class Response {
 
-	private HttpExchange exchange;
+	private final HttpExchange exchange;
 
 	public Response(HttpExchange exchange) {
 		this.exchange = exchange;
@@ -23,5 +20,6 @@ public abstract class Response {
 		outputStream.flush();
 		outputStream.close();
 	}
+
 	protected abstract int getHttpStatusCode();
 }
