+++ app/src/main/java/com/codesoom/assignment/response/ResponseCreated.java
@@ -2,16 +2,16 @@ package com.codesoom.assignment.response;
 
 import com.sun.net.httpserver.HttpExchange;
 
-public class ResponseCreated extends Response{
-    private static final int Created = 201;
+public class ResponseCreated extends Response {
+	private static final int Created = 201;
 
-    public ResponseCreated(HttpExchange exchange) {
-        super(exchange);
-    }
+	public ResponseCreated(HttpExchange exchange) {
+		super(exchange);
+	}
 
-    @Override
-    protected int getHttpStatusCode() {
-        return Created;
-    }
+	@Override
+	protected int getHttpStatusCode() {
+		return Created;
+	}
 }
