+++ app/src/main/java/com/codesoom/assignment/response/ResponseNoContent.java
@@ -3,14 +3,14 @@ package com.codesoom.assignment.response;
 import com.sun.net.httpserver.HttpExchange;
 
 public class ResponseNoContent extends Response {
-    private static final int NO_CONTENT = 204;
+	private static final int NO_CONTENT = 204;
 
-    public ResponseNoContent(HttpExchange exchange) {
-        super(exchange);
-    }
+	public ResponseNoContent(HttpExchange exchange) {
+		super(exchange);
+	}
 
-    @Override
-    protected int getHttpStatusCode() {
-        return NO_CONTENT;
-    }
+	@Override
+	protected int getHttpStatusCode() {
+		return NO_CONTENT;
+	}
 }
